Add files via upload
1Q:
Description: This code defines an Employee class and a subclass called FullTimeEmployee to model employees in a company. The Employee class has a class variable emp_count that keeps track of the total number of employees. Its constructor initializes the employee's name, family, salary, and department, and increments the employee count each time a new employee object is created. The class also contains a method avg_salary that calculates the average salary from a list of employees by summing their salaries and dividing by the total number of employees. The FullTimeEmployee class inherits from the Employee class, and its constructor simply calls the parent constructor to initialize the necessary attributes. The program creates a list of employees, including both Employee and FullTimeEmployee objects, and calculates the average salary using the avg_salary method from both the Employee and FullTimeEmployee classes, printing the results.

2.Q: Description: This code generates a random vector of size 20 with values between 1 and 20, and then reshapes it into a 4x5 array. The task is to replace the maximum element in each row with the value 1. First, the array before the replacement is printed. Then, the code calculates the maximum value of each row using np.max with axis=1, reshapes it, and replaces those maximum values in the original array with 1. Finally, the array after the replacement is printed.